Updated for writing to logging.txt
This version writes a log file logging.txt if the receiving callsign is in the decoded message. 
So, if you start rtlsdr_wsprd with your own callsign, it will write a log entry to logging.txt every time it decodes a message with your callsign in the message. I used this for testing multiple different transmitters.